requirements.txt
selenium==4.16.0
python-dateutil==2.8.2
fake-useragent==1.4.0
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
backoff==2.2.1
webdriver-manager==4.0.1
openai==0.28.0
streamlit==1.29.0
markdown==3.5.1
